+++ jpostman-annotations/src/main/java/io/jpostman/annotations/runtime/PreparedContext.java
@@ -1,13 +1,17 @@
 package io.jpostman.annotations.runtime;
 
 import io.jpostman.Collection;
+import io.jpostman.JPostman.Context;
 
 final class PreparedContext<C> {
+
 	final C context;
+	final Context loaded;
 	final Collection collection;
 
-	public PreparedContext(C context, Collection collection) {
+	PreparedContext(C context, Context loaded) {
 		this.context = context;
-		this.collection = collection;
+		this.loaded = loaded;
+		this.collection = loaded == null ? null : loaded.getCollection();
 	}
 }